Debug: output data6

+++ src/omv/modules/py_tf.c
@@ -326,9 +326,9 @@ STATIC void py_tf_regression_output_callback(void *callback_data,
         for (int i = 0, ii = params->output_channels; i < ii; i++) {
 
             mp_float_t temp = (float) (((float *) model_output)[i]);
-            printf( "raw output %.1f \t", temp);
+            printf( "raw output %.1f \t", (double)temp);
             mp_float_t up_temp = (float) ( (temp - params->output_zero_point) * params->output_scale);
-            printf( "scaled zeroed output %.1f \t", up_temp);
+            printf( "scaled zeroed output %.1f \t", (double)up_temp);
 
             ((mp_obj_list_t *) arg->out)->items[i] =
                     mp_obj_new_float( ((float) (((int8_t *) model_output)[i] - params->output_zero_point)) * params->output_scale);
